Add ability to use literal value on InputListLOC component; DRY tests
