+++ __tests__/components/editor/property/InputLookupQA.test.js
@@ -3,7 +3,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 import InputLookupQA from 'components/editor/property/InputLookupQA'
-import { renderMenuFunc, renderTokenFunc } from 'components/editor/property/renderTypeaheadFunctions'
 
 const plProps = {
   id: 'lookupComponent',
@@ -37,66 +36,6 @@ const plProps = {
   isLoading: false,
   search: jest.fn(),
 }
-
-const p2Props = {
-  id: 'lookupComponent',
-  propertyTemplate: {
-    mandatory: 'false',
-    repeatable: 'true',
-    type: 'lookup',
-    resourceTemplates: [],
-    valueConstraint: {
-      valueTemplateRefs: [],
-      useValuesFrom: [
-        'lookupQaLocNames',
-        'lookupQaLocSubjects',
-      ],
-      valueDataType: {
-        dataTypeURI: 'http://id.loc.gov/ontologies/bibframe/Agent',
-      },
-      defaults: [],
-    },
-    propertyURI: 'http://id.loc.gov/ontologies/bflc/target',
-    propertyLabel: 'Name Lookup',
-  },
-  lookupConfig: [
-    {
-      label: 'LOC person [names] (QA)',
-      uri: 'urn:ld4p:qa:names:person',
-      authority: 'locnames_ld4l_cache',
-      subauthority: 'person',
-      language: 'en',
-      component: 'lookup',
-    },
-    {
-      label: 'LOC all subjects (QA)',
-      uri: 'urn:ld4p:qa:subjects',
-      authority: 'locsubjects_ld4l_cache',
-      subauthority: '',
-      language: 'en',
-      component: 'lookup',
-    },
-  ],
-  isLoading: false,
-  search: jest.fn(),
-}
-
-const multipleResults = [
-  { authLabel: 'Person', authURI: 'PersonURI' },
-  { uri: 'http://id.loc.gov/authorities/names/n860600181234', label: 'Names, Someone' },
-  { authLabel: 'Subject', authURI: 'SubjectURI' },
-  { uri: 'http://id.loc.gov/authorities/subjects/sh00001861123', label: 'A Specific Place' },
-]
-
-const validNewURIResults = [{
-  customOption: true,
-  label: 'http://id.loc.gov/authorities/subjects/123456789',
-}]
-
-const validNewLiteralResults = [{
-  customOption: true,
-  label: 'Some non URI string',
-}]
 
 describe('<InputLookupQA />', () => {
   const mockFormDataFn = jest.fn()
@@ -136,81 +75,6 @@ describe('<InputLookupQA />', () => {
     wrapper.find('#lookupComponent').simulate('change')
 
     expect(mockFormDataFn).toHaveBeenCalled()
-  })
-
-  it('links the tokens when there is a URI', () => {
-    const option = {
-      uri: 'http://id.loc.gov/authorities/names/no2017003958',
-      id: 'no2017003958',
-      label: 'Ju, Peijian',
-    }
-
-    const tokenWrapper = shallow(renderTokenFunc(option, { labelKey: 'label' }, 0))
-    expect(tokenWrapper.exists('a[href="http://id.loc.gov/authorities/names/no2017003958"]')).toEqual(true)
-  })
-
-  it('does not link the tokens when there is no URI', () => {
-    const option = {
-      id: 'no2017003958',
-      label: 'Ju, Peijian',
-    }
-
-    const tokenWrapper = shallow(renderTokenFunc(option, { labelKey: 'label' }, 0))
-    expect(tokenWrapper.exists('a')).toEqual(false)
-  })
-
-  it('shows menu headers with lookup source labels and values in the dropdown when provided results', () => {
-    const menuWrapper = shallow(renderMenuFunc(multipleResults, p2Props))
-    const menuChildrenNumber = menuWrapper.children().length
-    // One top level menu component
-
-    expect(menuWrapper.find('ul').length).toEqual(1)
-    // Four children, with two headings and two items
-    expect(menuChildrenNumber).toEqual(4)
-    expect(menuWrapper.childAt(0).html()).toEqual('<li class="dropdown-header">Person</li>')
-    expect(menuWrapper.childAt(1).childAt(0).text()).toEqual('Names, Someone')
-    expect(menuWrapper.childAt(2).html()).toEqual('<li class="dropdown-header">Subject</li>')
-    expect(menuWrapper.childAt(3).childAt(0).text()).toEqual('A Specific Place')
-  })
-
-  it('shows a single new valid URI value with the correct header when no other matches are found', () => {
-    const menuWrapper = shallow(renderMenuFunc(validNewURIResults, p2Props))
-    const menuChildrenNumber = menuWrapper.children().length
-    // One top level menu component
-
-    expect(menuWrapper.find('ul').length).toEqual(1)
-    // Two children, with one headings and one custom item
-    expect(menuChildrenNumber).toEqual(2)
-    expect(menuWrapper.childAt(0).html()).toEqual('<li class="dropdown-header">New URI</li>')
-    expect(menuWrapper.childAt(1).childAt(0).text()).toEqual('http://id.loc.gov/authorities/subjects/123456789')
-  })
-
-  it('does show a single new literal value when no other matches are found', () => {
-    const menuWrapper = shallow(renderMenuFunc(validNewLiteralResults, p2Props))
-    const menuChildrenNumber = menuWrapper.children().length
-    // One top level menu component
-
-    expect(menuWrapper.find('ul').length).toEqual(1)
-    // Nothing shown because the entered URI is not valid
-    expect(menuChildrenNumber).toEqual(2)
-    expect(menuWrapper.childAt(0).html()).toEqual('<li class="dropdown-header">New Literal</li>')
-    expect(menuWrapper.childAt(1).childAt(0).text()).toEqual('Some non URI string')
-  })
-
-  it('shows menu headers for both lookups and new valid URI value with the correct headers when matches are found', () => {
-    const menuWrapper = shallow(renderMenuFunc(multipleResults.concat(validNewURIResults), p2Props))
-    const menuChildrenNumber = menuWrapper.children().length
-    // One top level menu component
-
-    expect(menuWrapper.find('ul').length).toEqual(1)
-    // Five children, with three headings and three items
-    expect(menuChildrenNumber).toEqual(6)
-    expect(menuWrapper.childAt(0).html()).toEqual('<li class="dropdown-header">Person</li>')
-    expect(menuWrapper.childAt(1).childAt(0).text()).toEqual('Names, Someone')
-    expect(menuWrapper.childAt(2).html()).toEqual('<li class="dropdown-header">Subject</li>')
-    expect(menuWrapper.childAt(3).childAt(0).text()).toEqual('A Specific Place')
-    expect(menuWrapper.childAt(4).html()).toEqual('<li class="dropdown-header">New URI</li>')
-    expect(menuWrapper.childAt(5).childAt(0).text()).toEqual('http://id.loc.gov/authorities/subjects/123456789')
   })
 
   describe('Errors', () => {
